+++ src/Comparison/Comparator.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Version\Comparator;
+namespace Version\Comparison;
 
 use Version\Version;
 
-interface ComparatorInterface
+interface Comparator
 {
     public function compare(Version $version1, Version $version2): int;
 }