+++ tests/Comparison/Constraint/OperatorConstraintParsingTest.php
@@ -2,23 +2,23 @@
 
 declare(strict_types=1);
 
-namespace Version\Tests\Constraint;
+namespace Version\Tests\Comparison\Constraint;
 
 use PHPUnit\Framework\TestCase;
-use Version\Constraint\ComparisonConstraint;
-use Version\Constraint\CompositeConstraint;
-use Version\Exception\InvalidComparisonConstraintStringException;
+use Version\Comparison\Constraint\OperatorConstraint;
+use Version\Comparison\Constraint\CompositeConstraint;
+use Version\Exception\InvalidConstraintStringException;
 
-class ComparisonConstraintParsingTest extends TestCase
+class OperatorConstraintParsingTest extends TestCase
 {
     /**
      * @test
      */
     public function it_parses_simple_constraint(): void
     {
-        $constraint = ComparisonConstraint::fromString('>=1.2.0');
+        $constraint = OperatorConstraint::fromString('>=1.2.0');
 
-        $this->assertInstanceOf(ComparisonConstraint::class, $constraint);
+        $this->assertInstanceOf(OperatorConstraint::class, $constraint);
         $this->assertSame('>=', $constraint->getOperator());
         $this->assertSame('1.2.0', (string) $constraint->getOperand());
     }
@@ -28,9 +28,9 @@ class ComparisonConstraintParsingTest extends TestCase
      */
     public function it_assumes_equals_as_default_operator_if_operator_not_supplied(): void
     {
-        $constraint = ComparisonConstraint::fromString('1.2.0');
+        $constraint = OperatorConstraint::fromString('1.2.0');
 
-        $this->assertInstanceOf(ComparisonConstraint::class, $constraint);
+        $this->assertInstanceOf(OperatorConstraint::class, $constraint);
         $this->assertSame('=', $constraint->getOperator());
         $this->assertSame('1.2.0', (string) $constraint->getOperand());
     }
@@ -40,7 +40,7 @@ class ComparisonConstraintParsingTest extends TestCase
      */
     public function it_parses_range_constraint(): void
     {
-        $constraint = ComparisonConstraint::fromString('>=1.2.3 <1.3.0');
+        $constraint = OperatorConstraint::fromString('>=1.2.3 <1.3.0');
 
         $this->assertInstanceOf(CompositeConstraint::class, $constraint);
         $this->assertCompositeConstraintIsIdentical(
@@ -58,7 +58,7 @@ class ComparisonConstraintParsingTest extends TestCase
      */
     public function it_parses_composite_constraints_containing_logical_operators(): void
     {
-        $constraint = ComparisonConstraint::fromString('>=1.0.0 <1.1.0 || >=1.2.0');
+        $constraint = OperatorConstraint::fromString('>=1.0.0 <1.1.0 || >=1.2.0');
 
         $this->assertInstanceOf(CompositeConstraint::class, $constraint);
         $this->assertCompositeConstraintIsIdentical(
@@ -88,10 +88,10 @@ class ComparisonConstraintParsingTest extends TestCase
     public function it_raises_exception_if_the_constraint_string_is_empty(): void
     {
         try {
-            ComparisonConstraint::fromString('  ');
+            OperatorConstraint::fromString('  ');
 
             $this->fail('Exception should have been raised');
-        } catch (InvalidComparisonConstraintStringException $ex) {
+        } catch (InvalidConstraintStringException $ex) {
             $this->assertSame('Constraint string must not be empty', $ex->getMessage());
         }
     }
@@ -102,10 +102,10 @@ class ComparisonConstraintParsingTest extends TestCase
     public function it_raises_exception_if_the_constraint_string_cannot_be_parsed(): void
     {
         try {
-            ComparisonConstraint::fromString('invalid');
+            OperatorConstraint::fromString('invalid');
 
             $this->fail('Exception should have been raised');
-        } catch (InvalidComparisonConstraintStringException $ex) {
+        } catch (InvalidConstraintStringException $ex) {
             $this->assertSame("Constraint string: 'invalid' seems to be invalid and it cannot be parsed", $ex->getMessage());
         }
     }
@@ -116,10 +116,10 @@ class ComparisonConstraintParsingTest extends TestCase
     public function it_raises_exception_if_the_constraint_contains_operator_that_cannot_be_parsed(): void
     {
         try {
-            ComparisonConstraint::fromString('"100');
+            OperatorConstraint::fromString('"100');
 
             $this->fail('Exception should have been raised');
-        } catch (InvalidComparisonConstraintStringException $ex) {
+        } catch (InvalidConstraintStringException $ex) {
             $this->assertSame("Constraint string: '\"100' seems to be invalid and it cannot be parsed", $ex->getMessage());
         }
     }
@@ -130,10 +130,10 @@ class ComparisonConstraintParsingTest extends TestCase
     public function it_raises_exception_if_the_constraint_contains_version_that_cannot_be_parsed(): void
     {
         try {
-            ComparisonConstraint::fromString('>100');
+            OperatorConstraint::fromString('>100');
 
             $this->fail('Exception should have been raised');
-        } catch (InvalidComparisonConstraintStringException $ex) {
+        } catch (InvalidConstraintStringException $ex) {
             $this->assertSame("Constraint string: '>100' seems to be invalid and it cannot be parsed", $ex->getMessage());
         }
     }
@@ -144,10 +144,10 @@ class ComparisonConstraintParsingTest extends TestCase
     public function it_raises_exception_if_the_constraint_contains_invalid_logical_operation(): void
     {
         try {
-            ComparisonConstraint::fromString('>=1.0.0 <1.1.0 ||');
+            OperatorConstraint::fromString('>=1.0.0 <1.1.0 ||');
 
             $this->fail('Exception should have been raised');
-        } catch (InvalidComparisonConstraintStringException $ex) {
+        } catch (InvalidConstraintStringException $ex) {
             $this->assertSame("Constraint string: '>=1.0.0 <1.1.0 ||' seems to be invalid and it cannot be parsed", $ex->getMessage());
         }
     }
@@ -157,7 +157,7 @@ class ComparisonConstraintParsingTest extends TestCase
         self::assertSame($expectedOperator, $compositeConstraint->getOperator());
 
         foreach ($compositeConstraint->getConstraints() as $i => $constraint) {
-            /* @var $constraint \Version\Constraint\ComparisonConstraint */
+            /* @var $constraint \Version\Comparison\Constraint\OperatorConstraint */
 
             if ($constraint instanceof CompositeConstraint) {
                 self::assertCompositeConstraintIsIdentical(